requirements.txt
requests==2.22.0
numpy==1.15.1
flask==1.1.1
pandas==0.23.4
pyyaml==3.11
googlefinance